Forced to make __audioFiles in cli.CLIArguments a List<String> in order for JCommander to recognize it as the main parameter. Added modified formatter class for use by JCommander.usage() Rearranged POM Added in a class to handle property file loading Implemented type detection using SimpleMagic library Implemented semi-intelligent audio player selection; all concrete implementations of AbstractAudioPlayer will now need a static block with contents similar to those found in the included players. All concrete implementations of AbstractAudioPlayer will now need to be declared in props/playerHandlers.properties, in the form of "key=fully_qualified_class_name". The value of key does not matter, as long as it is unique. Created PlayerTypeMap.java to facilitate player selection.
Changes to exception handling and logging in numerous places.
Prefixed many instance method calls and instance fields with "this."
Minor cleanup, reformatting, and refactoring.

Removed log4j2-test.properties, because that wasn't supposed to be part of the repo in the first place. 🤦‍♂️
Regression: At some point I broke file glob parameters, and I don't know how long it's been that way because I hadn't been testing for it. 😅